refactor: 스타일 수정 & 오타 수정
