+++ src/pages/PostPage/index.ts
@@ -1,0 +1,1 @@
+export { default as PostPage } from './PostPage';
